requirements.txt
cassandra-driver
PyYAML
